refactor(ScoreBoard) : 승/패 계산 책임을 BattleResultCalculator로 전가
